+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Fix box-box collision by computing the contact normal at the closest point of approach instead of at the center of
   the source box ([GH-839](https://github.com/NVIDIA/warp/pull/839)).
 - Fix potential uninitialized memory issues in `wp.tile_sort()` when an array of `NaN` values were provided ([GH-836](https://github.com/NVIDIA/warp/issues/836).
+- Fix `wp.tile_min()` and `wp.tile_argmin()` for large tiles with low occupancy ([GH-725](https://github.com/NVIDIA/warp/issues/725)).
 
 ## [1.8.0] - 2025-07-01